+++ src/search-commons/model/records/baseobjectdetectionimageobject-record.ts
@@ -8,6 +8,7 @@ export interface BaseObjectDetectionImageObjectRecordType extends BaseImageRecor
 
 export class BaseObjectDetectionImageObjectRecord extends BaseImageRecord implements BaseObjectDetectionImageObjectRecordType {
     static objectDetectionImageObjectFields = {
+        category: new BaseEntityRecordFieldConfig(GenericValidatorDatatypes.NAME, new NameValidationRule(true)),
         detector: new BaseEntityRecordFieldConfig(GenericValidatorDatatypes.NAME, new NameValidationRule(true)),
         key: new BaseEntityRecordFieldConfig(GenericValidatorDatatypes.NAME, new NameValidationRule(true)),
         keySuggestion: new BaseEntityRecordFieldConfig(GenericValidatorDatatypes.NAME, new NameValidationRule(false)),
@@ -22,6 +23,7 @@ export class BaseObjectDetectionImageObjectRecord extends BaseImageRecord implem
         precision: new BaseEntityRecordFieldConfig(GenericValidatorDatatypes.NUMBER, new NumberValidationRule(false, 0, 999999, 0))
     };
 
+    category: string;
     detector: string;
     key: string;
     keySuggestion: string;
@@ -38,7 +40,8 @@ export class BaseObjectDetectionImageObjectRecord extends BaseImageRecord implem
     toString() {
         return 'BaseObjectDetectionImageObjectRecord Record {\n' +
             '  id: ' + this.id + ',\n' +
-            '  class: ' + this.detector + ',\n' +
+            '  detector: ' + this.detector + ',\n' +
+            '  category: ' + this.category + ',\n' +
             '  key: ' + this.key + ',\n' +
             '  pos: ' + this.objX + ',' + this.objY + '(' + this.objWidth + ',' + this.objHeight + ')\n' +
             '}';