+++ src/search-commons/model/container/facets.ts
@@ -1,3 +1,4 @@
+export type FacetValueType = 'number' | 'date' | 'string';
 export class Facet {
     public facet: Array<Array<any>> = [];
     public selectLimit?: number;
@@ -6,3 +7,19 @@ export class Facet {
 export class Facets {
     public facets: Map<string, Facet> = new Map<string, Facet>();
 }
+
+export class FacetUtils {
+    public static calcFacetValueType(valueType: FacetValueType): FacetValueType {
+        if (valueType === undefined || valueType === null) {
+            return 'string'
+        }
+
+        return valueType;
+    }
+
+    public static generateFacetCacheKey(tableKey: string, facetKey: string): string {
+        return tableKey + '_' + facetKey;
+    }
+}
+
+