+++ docs/CHANGELOG.md
@@ -28,6 +28,7 @@
 - common: improved null-handling for sql-filter
 - common: improved beanutils to get array-values too
 - common: added flag for special facets to read only if explicitly named
+- common: added debug-messages for sql
 
 ### bug fixes
 - none